Share start option merging across the regex machinery
